Avoid buffer overflow in expandString()
The test in the old code (formatCursor >= MAX_IMAGE_PATH_LENGTH - 1)
only ensured there was room for one more character (plus the
NUL-terminator) in any iteration of the loop and thus the expansion
of '%p', '%d' or '%t' might not fit.

This was highlighted by a warning issued by gcc 9.

Signed-off-by: Keith W. Campbell <keithc@ca.ibm.com>