Optimize range based for loops
In some parts of the JITServer code we use range-based for loops to
iterate over collections. The typical expression is:
  for (auto it : collection)
This form accesses the elements of the collection by value. If the
element of the collection is a complex data structure involving
other STL data structures (like maps/sets, etc) they will be copy
constructed adding extra overhead.
A better solution is to access the elements by reference like this:
  for (auto& it : collection)

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>