Fix serialization of entire CHTable
The JITClient needs to send the entire CHTable to the JITServer during
the first compilation when connecting to a new JITServer. Currently
the hierarchy is collected starting from the java/lang/Object class
and recursively finding all the subclasses. However, this approach
does not catch the interfaces. In the end all is well due to a quirk
of the implementation that sends CHTable updates, but it's better to
do the right thing and correctly send the entire CHTable when the
JITServer asks for it.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>