Ignore -XcompilationThreads0 option
-XcompilationThreads<N> option can be used to specify the desired number
of compilation threads. The maximum valid number is 7 and the JIT will
silently ignore the option if a larger value is specified. If the user
specifies a value of 0, then the JIT will trigger a FATAL assert.
This commit modifies the code so that a value of 0 will too be silently
ignored, in which case the JIT will use its internal heuristics to determine
how many compilation threads to create.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>