Fix memory leak in JITServer related to IProfiler
JITServerIProfiler::setCallCount uses two types of caches to reduce
the number of messages sent to the client: a per-client cache which
uses persistent memory and a per-compilation cache which uses scratch
memory. The memory leak happens because we use persistent memory to
store an entry in the per-compilation cache. This cache will vanish
when the compilation is over because all of the backing scratch
memory segments are returned to the OS but the attached entries
allocated with persistent memory will leak.
The fix is to use scratch memory for the per-compilation cache and
persistent memory for the per-client cache.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>