Fix memory leak due to caching of resolved methods
The JITServer keeps a cache with information that ease the creation
of new resolved methods. This information is sent by the JITClient
and includes three std::string blobs representing the bytes for
PersistentJittedBodyInfo, PersistentMethodInfo and a serialized version
of the IProfiler information related to the method in question.
Caching these three strings is wrong because they come directly from
the protobuf infrastructure which uses the system heap for memory
allocations. So even though our cache is created with scratch memory
the underlying buffers for these strings still use system heap memory.
These buffers will leak because we don't explicitly free memory, but
rather rely on freeing the entire backing scratch memory segments.

The solution adopted in this commit is to cache pointers to memory
allocated from the scratch segments with `allocateHeapMemory` and then
to recreate the strings when retrieving the required information from
the cache. Further efficiency can be achieved by not having to recreate
those strings, but doing so requires some code restructuring which is
better left for a separate PR.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>