Add structral comparison barriers
In #7743 a flag (J9ClassCanSupportFastSubstitutability) is added to each
valueType class if it does not contain any floating point primitives or
references that might potentially contain an unflattened valuetype or a
flattened valuetype that does not contain
J9ClassCanSupportFastSubstitutability.

For these classes a structural comparison can be performed. This is
essentially performing a memory comparison over the length of the
instances, if they are equivalent the result is true, otherwise false.
Since it will only contain scalar primitives and (non value) references,
equality can be determined by simply performing a memory comparison.
Unlike with floating point types or non-flattened valuetypes where
determining equality is more complicated.

A barrier helper is required as these instances may contain references.
In modes where no read barriers are required memcmp can be used.

Signed-off-by: Tobi Ajila <atobia@ca.ibm.com>