Disable garbage collection for value types test
The garbage collector does not currently have support for value types.
It is only by chance that all tests are currently working. The initial
work to support value types will start causing the GC to runtime assert
when the policy does not support value types. As support for more GC
policies is added, we can update the policy used for this test.

Signed-off-by: Andrew Young <youngar17@gmail.com>