Add missing barriers to jProfiling helpers
Previously, the tryLock and unlock methods on the jProfiling hash table
implementation did not contain any read or write barriers. While this is
fine on architectures such as x86-64, this causes problems with the weak
memory model used in Power. Failing to include the barriers could allow
reordering of memory accesses to cause incorrect behaviour with
unpredictable results.

Signed-off-by: Ben Thomas <ben@benthomas.ca>