Add JITServer-specific functional tests
This patch adds testing for a few basic JITServer scenarios.

The tests only excercise the compiler, not the compiled code (i.e. the
compiled code is never executed). The tests are somewhat portable to
other *nix platforms, but not Windows, and have only been validated
on Linux. A runtime assert makes sure to check that the tests only
execute on Linux for now.

Signed-off-by: Younes Manton <ymanton@ca.ibm.com>