Fix String compressible API loop entry conditions
The loop entry conditions should be traversing the input array starting
from the given offset for the given length number of characters.
Previously passing a starting offset which is greater than the length
would result in the API never entering the loop and always returning
true, even though the input may have contained non-compressible
characters.

We also take this opportunity to rename `compressible` to a better name
which represents what the API is doing, i.e. `canEncodeAsLatin1`.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>