Verify methods expected to be native are, across depedence relations
Defined a new struct, W, to describe a dependency relationship of
a recognized method upon other recognized methods, arrays to aggregate
those relationships on a per class basis, and a new method,
areNativeDependenciesSatisfied, to navigate through those relationships.

areNativeDependenciesSatisfied checks for the methods upon which the
implementation of a potentially recognized method depends.  Any of those
methods upon which it depends that are expected to be native are checked
for whether they really are native.  If such a method is expected to be
native, but is not or cannot be located, the method that depends upon it
not treated as a recognized method.  These dependencies are checked
recursively.

As this is a prototype, the arrays of dependecies are incomplete, and
only exist for the java.util.concurrent.atomic.AtomicInteger,
sun.misc.Unsafe and jdk.internal.misc.Unsafe classes.

Also added code that is conditionally compiled out by default to dump
the contents of recognized method tables as XML.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>