Update zlib to 1.2.11
zlib 1.2.3 has known security vulnerabilities
[CVE-2016-9840](https://nvd.nist.gov/vuln/detail/CVE-2016-9840),
[CVE-2016-9841](https://nvd.nist.gov/vuln/detail/CVE-2016-9841),
[CVE-2016-9842](https://nvd.nist.gov/vuln/detail/CVE-2016-9842),
[CVE-2016-9843](https://nvd.nist.gov/vuln/detail/CVE-2016-9843).

Apply a subset of the openj9 changes from the zlib 1.2.3 version to the
latest 1.2.11 version. Discarded changes that are not required in order
to minimize changes to the original zlib code.

Moves zconf.h to the include directory. zlib.h is in the include
directory and includes this file so it should be in the same place.

zlib 1.2.11 source https://www.zlib.net/zlib-1.2.11.tar.gz
zlib 1.2.3 source https://www.zlib.net/fossils/zlib-1.2.3.tar.gz

Signed-off-by: Peter Shipton <Peter_Shipton@ca.ibm.com>