Add comment on where _recompilationMethodInfo is freed
`CompilationInfoPerThreadRemote::_recompilationMethodInfo` was allocated
in the persistent memory in `CompilationInfoPerThreadRemote::processEntry()`.
`_recompilationMethodInfo` is related to per compilation. Ideally it should be
allocated in the scratch memory. However the scratch memory is part of
the compilation object which has not been created when the entry is being processed.

During the compilation, `_recompilationMethodInfo` is passed
to `J9::Recompilation::_methodInfo`, which will get freed in `populateBodyInfo()`
when creating the method meta data during the compilation.
Since `_recompilationMethodInfo` is already freed in `populateBodyInfo()`,
no need to free it at the end of the compilation in `CompilationInfoPerThreadRemote::processEntry()`.

Closes #7988

Signed-off-by: a7ehuo <Annabelle.Huo@ibm.com>