Set JITServer Compilations Options on the client instead
Previously, these options were being set as global options
on the server side during startup. This is incorrect, as options
set on the server are overwritten by the client options when the
server receives a compilation request from a client.

This commit now sets the options on the client, right before a remote
compilation request is sent to the server.

Signed-off-by: Dhruv Chopra <Dhruv.C.Chopra@ibm.com>