Fix data race when locking JProfiling tables
When adding new values to a JProfiling table, it is necessary for the
JIT helpers to lock the table in question to avoid conflicts. However,
the previous method of locking involved a data race on the _metaData
field. Specifically, the read of _metaData to determine the value to
CAS was not atomic. The result was that xlc compiled this into two
reads, one of _metaData.otherIndex and one of _metaData.rawData. If
another thread were to release the lock in between these two reads, this
could result in the value swapped into _metaData having otherIndex < 0,
which would trigger the "JIT access not disabled" assert.

This has been corrected by performing a single volatile read of
_metaData to determine what value to use. While it's not clear that this
eliminates all undefined behaviour here, it does fix the immediate
problem.

Signed-off-by: Ben Thomas <ben@benthomas.ca>