Fix initialization of J9::Compilation::_isOutOfProcessCompilation
The field `J9::Compilation::_isOutOfProcessCompilation` is used by the compiler
to determine whether this is a compilation performed at the JITServer. The
problem is that during the construction of the compiler object many other
objects are instantiated and during these operations we enquire the status
of the _isOutOfProcessCompilation field which might be unset. According to the
C++ standard fields are initialized in the order of their declaration. In order
to have `_isOutOfProcessCompilation` initialized before anything else we might
need to declare it at the omr level as the first field in OMR::Compilation and
to pass some information about being an out-of-process compilation to the
constructor of J9::Compilation.

This commit implements a simpler solution: since the JITServer can perform only
out-of-process compilations and nothing else, we can make `_isOutOfProcessCompilation`
a static field that is set once as soon as we learn that this process is a
JITServer.

Fixes #8150

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>