Use new OMR API to implement getSystemCpuLoad
A new OMR port library API `omrsysinfo_get_CPU_load` has been
implemented which calculates the CPU load for us. The OMR API is a
replica of how we currently calculate the CPU load (the logic from
OpenJ9 has been sunk down into OMR).

In this commit we make use of the new API as it provides support for
all operating systems we build OpenJ9 on.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>