Increase JITServer default timeout value (0.18.0)
The current timeout value for JITServer is 2000 ms (the same as
for the client JVM) which is too low. When the client is using
`-Xjit:enableJITServerHeuristics`  cheap compilations (those
performed at 'cold' and 'no-opt' optimization levels) are
performed locally in order to avoid network latency. If there is
a large stretch of such cheap compilations, the server may not
'hear' from the client for a long time and close the connection.
To avoid this scenario a larger timeout value at the server needs
to be used.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>