JIT Support for getCallerClass in recent JDK versions
In recent versions of the JDK the getCallerClass taking an int argument has
been dropped with an argument-less version remaining. The behaviour of this
API is equivalent to getCallerClass with an argument of -1. This change adds
support for this second version of getCallerClass and for the renamed package
of the class this method is found in.

Signed-off-by: Andrew Craik <ajcraik@ca.ibm.com>