Remove use of old VMX record form instruction names
Previously, the names of the VMX record form instructions in the
TR::InstOpCode::Mnemonic enum were inconsistent with the names of other
record form instructions. As a result, these instructions were renamed
in OMR and aliases were temporarily left for the old names to avoid
build breakages. In order to allow these aliases to be removed from OMR,
references to the old names have been changed to use the new names.

Signed-off-by: Ben Thomas <ben@benthomas.ca>