Retrieve constant dynamic info from the JITClient
Dynamic constant is not stored at the JITServer.
The server will send one message to retrieve both the pointer
and the object values in `TR_ResolvedJ9JITServerMethod::dynamicConstant()`.
It has dependency on the new `dynamicConstant` API
defined in eclipse/omr#4726.

Also increased the JITServer `MINOR_NUMBER` to `3` since
the new message types have to be handled at the client as well.

This commit fixes the server crash when running
Java 11 sanity test `CondyPrimitive_1`.

Related to #7567

Signed-off-by: Annabelle Huo <Annabelle.Huo@ibm.com>