Add walkReferenceChain JITServer Support
The purpose of this change is to avoid using object addresses on JITServer.
Using object addresses on JITServer creates problems when GC changes the object addresses. Instead, JITServer must use some kind of handle (a GlobalReference or some other GC root) to hold the object so that JITServer always has the valid object addresses.

walkReferenceChain is special in that it needs half of its information from the JITServer(TR::Node) and half of its information from JITClient(object address). During the walk, we need vmaccess so that GC doesn't move the object of interest. We can only have vmaccess on JITClient, which means we need to do walkReferenceChain on JITClient.

This change makes JITServer to generate the offsets information JITClient needed for the walk and passes to JITClient in the form of a vector.

Issue: #8109

Signed-off-by: Harry Yu <harryyu1994@gmail.com>