Fix incorrect profiling of arraycopy length
Under some circumstances, it is possible for a call to Object.clone on
an object which has a known array type to be turned into an allocation
of a new array followed by an arraycopy. When this occurs, the resulting
arraycopy node will have the same bytecode information as the call node
it was created from.

While this is generally fine, it can create issues later when we attempt
to read profiling information for the arraycopy node, which we expect to
contain the profiled length of the arraycopy. However, the IProfiler
information that is read is actually related to the original call node,
containing profiled J9Class pointers. This can result in the arraycopy
node being specialized with a nonsensical (and possibly invalid) length.

This has been corrected by ensuring that addresses are being requested
before reading IProfiler value profiling information in response to a
request for profiling information. Instead, no profiling information
will be read, thus preventing the node from being specialized.

Signed-off-by: Ben Thomas <ben@benthomas.ca>