Fix out of range immediate for BNDCHK
As part of an upcoming refactor of the Power binary encoder, new asserts
are being added to ensure that immediates provided to instructions are
within the range that they should be. As it turns out, the BNDCHK
evaluator is capable of creating instructions with an out-of-range value
for their SI field if the bound being checked against is a negative
constant. While this may seem to be an impossible situation, it is
actually possible for escape analysis to create such nodes in an
unreachable code path by casting a newly allocated object to an array
type and then indexing into it. At runtime, this will always cause an
exception, but the code for the resulting BNDCHK is still generated and
will contain a dummy length which happens to be negative.

Signed-off-by: Ben Thomas <ben@benthomas.ca>