Fix out of range immediate for monexit
As part of an upcoming refactor of the Power binary encoder, new asserts
are being added to ensure that immediates provided to instructions are
within the range that they should be. As it turns out, the evaluator for
monexit creates an addi instruction with an out-of-range value for its
SI field due to unnecessarily ANDing the immediate with 0xffff. While
this would previously encode correctly, the new asserts will fire when
that immediate is negative, since it will appear to be an out-of-range
positive value.

Signed-off-by: Ben Thomas <ben@benthomas.ca>