Set estimate when encoding per-CC helper snippets
Previously, when performing binary encoding for the per-CC helper
snippets on Power, binary length estimates were not being set on the
instructions being encoded. While this was fine in the past, since these
estimates were not being used in this case, an impending refactor to the
Power binary encoder will be adding asserts to ensure that these values
were set correctly initially. As a result of this, it is now necessary
to ensure that estimateBinaryLength is called on an instruction prior to
calling generateBinaryEncoding on it.

Signed-off-by: Ben Thomas <ben@benthomas.ca>