Update AccessMode and SignatureType in VarHandle to support JEP-370
AddressVarHandleGenerator.addAccessModeTypeMethod accesses a field named
"at" in VarHandle.AccessMode enum. This field is currently named as
"signatureType". It is renamed from "signatureType" to "at" in order to
support JEP-370.

SignatureType enum values are renamed to work with
AddressVarHandleGenerator and VarForm:
1) getter -> GET
2) setter -> SET
3) compareAndSet -> COMPARE_AND_SET
4) compareAndExchange -> COMPARE_AND_EXCHANGE
5) getAndSet -> GET_AND_UPDATE
6) invalid removed since it is unused.

Also, enum SignatureType is renamed to AccessType in order to work with
VarForm and AddressVarHandleGenerator.

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>