Move setParameterLinkageRegisterIndex from OMR to OpenJ9
This function is very specific to a particular linkage as it associates
linkage registers for incoming arguments to parameters in the method.
Different linkages will do this in different ways so it does not make
sense to provide a default implementation in the base `OMRLinkage`
class.

Currently the XPLINK system linkage in OMR incorrectly defaults down to
executing the base implementation, which is correct for Java JIT
methods, but not correct for native XPLINK functions. We were
incorrectly mapping registers to parameters and were running into
troubles because this function was not properly implemented for XPLINK.

We avoid all of this by moving the implementation to OpenJ9 and making
the base function pure virtual to force the subclasses to properly
implement this API.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>