Improve performance on a full shared cache.
We see a significant slow down of startup speed when there are
many JVMs queuing for the write mutex of a full cache. It is not
necessary to acquire the write mutex in this case.

Return from j9shr_classStoreTransaction_start() after acquiring the
segment mutex but before acquiring the cache write mutex if the cache is
full.

Signed-off-by: Hang Shao <hangshao@ca.ibm.com>