Fix jniGetPrimitiveArrayCritical in the case of empty arraylets
Empty arraylets are treated as discontiguous arraylets. But since
they are empty there's no need to copy them in the critical section,
however, jniGetPrimitiveArrayCritical callers expects not NULL
values; therefore, in case of empty arraylets we advance the
pointer ahead of the header and return that address.

This fixes: https://github.com/eclipse/openj9/issues/8386

Signed-off-by: Igor Braga <higorb1@gmail.com>