Implement setParameterLinkageRegisterIndex pure virtual function
This pure virtual function as introduced in OMR to ensure all
subclasses of the linkage class implement it properly. Since
CHelperLinkage is only used for call-outs to CHelper APIs we cannot
implement this function properly, nor do we need to since no one should
ever be calling it. We add a fatal assert to ensure this is the case.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>