Initialize AccessMode.methodNameToAccessMode (HashMap)
AccessMode.methodNameToAccessMode is a HashMap<String, AccessMode>,
where the key is the method name and the value is the AccessMode enum
value. This hashmap is used in AccessMode.valueFromMethodName to
retrieve the AccessMode enum value for the corresponding method name. It
is also directly accessed from VarForm.linkFromStatic. It is needed so
that OpenJ9 can consume OpenJDK's VarForm class.

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>