Force stores to java_lang_ref_SoftReference_get to be non-volatile
Previously, special handling in the Il generator recognized stores to
the recognized field java_lang_ref_SoftReference_get reset the
volatility of the symbol for the store after the symbol was created.
Changing the volatility of a symbol after its creation significantly
complicates de-duplication for fabricated and created shadow symbols.
As an alternative, this commit changes that code so that the symbol is
created with the expected volatility, removing the need to change it
later on.

Signed-off-by: Leonardo Banderali <leonardo2718@protonmail.com>