Move double mapping after null leaf pointer has been initialized
Whenever the size of a discontiguous arraylet is a multiple of
region size in balanced GC, the GC creates an extra arraylet
leaf which contain zero bytes. In this case the arrayoid
corresponding to such leaf is set to NULL. If we double map
before the arrayed is set to NULL it is possible that such
arrayoid might contain garbage, which explains the assertion
failure (double map expects that an  arrayoid pointing to an
empty arraylet leaf to be NULL). Hence we move the double
mapping call to after setting the arrayoid to NULL.

Ran the test "CriticalRegionTest" which is the source of failures
1000 times here and all passed:
https://hyc-runtimes-jenkins.swg-devops.com/view/Test_grinder/job/Grinder/5871/

Fixes: https://github.com/eclipse/openj9/issues/8383

Signed-off-by: Igor Braga <higorb1@gmail.com>