Fix a JProfiling race condition on otherIndex
When adding a new key to a JProfiling value table, otherIndex was being
read before the table lock was acquired and then used after the lock was
acquired. This could result in erratic behaviour if otherIndex was
modified and then the lock was released between the time that otherIndex
was read and the lock was subsequently acquired. This has been corrected
and otherIndex will now be read after the lock is acquired to ensure it
remains consistent.

Signed-off-by: Ben Thomas <ben@benthomas.ca>