Handle jitNewValue in code generators
Because `newvalue` support is not implemented in the code generators,
the opcode is lowered to `new`. When lowering happens, the symbol on
the `new` is set to be the same as the one for `newvalue`, namely
jitNewValue.

As a result, `new` evaluators must check the node's symbol to determine
if the allocation is for a value type.

To avoid having to implement full support for inline value type
allocations, for now, we unconditionally call the JIT helper. While not
ideal for performance, it is a necessary step in incrementally
implementing value types.

Note that the AARCH64 code generator already calls whatever helper is
set on the `new` node unconditionally, so no special handling for
jitNewValue is required.

Signed-off-by: Leonardo Banderali <leonardo2718@protonmail.com>