Disable Power arraycopy inlining for very long lengths
As it turns out, the code for performing an arraycopy inline on Power is
currently broken for certain very long byte lengths. While this usually
doesn't matter, since the maximum length is bounded by the fact that
Java uses ints to represent array indices, it's possible for the
optimizer to create arraycopy nodes with very long lengths on
unreachable paths.

As part of the ongoing refactor of the Power binary encoder, a number of
new asserts are being added that catch any out of range immediates.
These asserts trigger on code generated for such arraycopy nodes even
though they're unreachable, since the code generated is nonsense.

This has been corrected by completely disabling arraycopy inlining on
Power for arraycopy nodes with lengths that do not fit in a 32-bit
integer.

Signed-off-by: Ben Thomas <ben@benthomas.ca>