Disable Power arraycopy inlining for all long lengths
Previously, the arraycopy inlining on Power was disabled for lengths
exceeding 256 bytes, but only in OMR's evaluator for 3-child arraycopy
nodes. However, as it turns out, our current inlining is suboptimal for
long lengths. To improve performance, this maximum length has now been
applied to the OpenJ9 ICF version of arraycopy inlining for 5-child
arraycopy nodes as well.

Signed-off-by: Ben Thomas <ben@benthomas.ca>