Prototype ILGen for defaultvalue bytecode
For a defaultvalue bytecode, if the class is resolved, generate a
newvalue instruction with default values for the fields as operands.
Recursively generate a newvalue for any fields that are themselves
value types.

The JIT compiler is not yet able to handle the case of an unresolved
class used in a defaultvalue bytecode instruction.  Simply fail the
compilation during ILGen if an unresolved class is encountered in that
situation.  The failure is reported as an
UnsupportedValueTypeOperation exception.  The intention is that these
can be replaced with OSR points in the future.

In order to track the frequency with which we are failing compilations
due to encountering unresolved classes during IL generation for the
defaultvalue bytecode, use static debug counters that begin with
the string "ilgen.abort/unresolved/defaultvalue".

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>