Branch to interface snippet should not have register dependencies
Register dependencies for interface dispatch are carefully constructed
such that they can be hung off the merge (or "done") label in the
mainline code.  They will cause incorrect register shuffling if they
are repeated on any other instruction in the dispatch sequence.  Ensure
that register dependencies created for this purpose only exist on
the merge label.

Signed-off-by: Daryl Maier <maier@ca.ibm.com>