Add support for multiple Artifactory servers
Artifactory bandwidth has become an issue in the CI builds.
This has caused issues with upload and download times, as well
as failing downloads due to infrastructure inadequacy.
Particularly for UNB machines, where a large portion of the
farm is located, combined with the pipe into UNB being small.

Being able to setup multiple Artifactory servers spread across
geographies and colocated with pools of machines will allow us
to push and more importantly pull from a nearby server.

The design leaves one site as the main (default) server (OSU). We
will always upload to the default server. We will only upload
to a secondary server(s) if there are machines with a matching
platform, colocated with another (non-default) server. The design
is laid out so it will scale with more servers without any code
change other than adding to the defaults.yml config.

Since we cannot determine where a test will land (geographically),
we need to upload to any/all servers that are colocated with machines
of matching platform. We also will always pass the default SDK URL.
We will add a curl wrapper on the nodes which have colocated servers.
This wrapper will redirect requests from one server to another. In
this case redirect OSU requests to UNB. We will also strip off
the user/password since a) The servers allow anonymous access and
b) The user's api key will be different for every server.

Changes introduced:
- Redesign the Artifactory config in defaults.yml to support multiple
  servers. Identify servers based on geo(graphy) and identify one
  geo as default.
- Redesign the Artifactory variables as a single hashmap. This keeps
  all the config together. Continue to write out the default server
  values to env, in order for the parent job to continue as-is.
- Move Artifactory setup onto the compile node. This is needed
  in order to determine where we are compiling and where we need
  to upload.
- Add support for uploading to a server behind a vpn. If we aren't
  compiling behind the vpn, stash the SDK for later when we can
  grab a node in the same geo as the server we need to push to.
  Note: We cannot publish buildInfo to vpn'd servers. Artifacts
  come from nodes but buildInfo comes from Master, so Master needs
  to see the server.
- Add some more details to the Artifacotry README. Also add more
  info specific to the UNB setup.

Also:
- Unrelated change to stop adding JAVADOC_LIB_URL to the CUSTOMIZED_SDK_URL.
  CUSTOMIZED_SDK_URL is for test and test doesn't need JAVADOC.

[skip ci]
Issue #8425

Signed-off-by: Adam Brousseau <adam.brousseau88@gmail.com>