Ensure that jitNewObject for value type is not evaluated inline
In the prototype implementation of value types, if the class specified
for the NEW bytecode instruction is a value type, an InstantiationError
must be thrown.  Value type instances are only created through the
defaultvalue and withfield bytecode instructions.

If the class passed to the jitNewObject helper is known to be a value
type, have canAllocateInlineClass return false for it.  That prevents
the JIT from generating inline code for allocating an instance,
including preventing it from being eligible for escape analysis.  That
ensures the helper is called at runtime, allowing it to throw the
required InstantiationError.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>