Fix bug in resolution of VT arrays and fields
When creating an array the elements of the array are set to the default
value. For arrays of regular identity types this is NULL. For arrays of
valueTypes this is the defaultValue of the valueType. As a result, if
the element type of the array is a valueType, the element type must be
fully initialized before the array can be created. This PR addresses
this problem.

Also, in the resolve field code, it is incorrectly fetching the
flattened field cache from the class where the fieldref is stored
instead of the class that defined the field. We havent noticed this
because the existing tests only fetches fields using getter methods.

Signed-off-by: Tobi Ajila <atobia@ca.ibm.com>