Add 'Q' prefix handling
Many places need proper handling for 'Q' prefix. In some places
it can be handled in the same way as 'L', whereas in some other
places, how to handle 'Q' needs to be further considered or
determined if the situation change (e.g. if 'Q' ends up being
passed by value).

Related to #12237

Co-authored-by:: Yiling Han <Yiling.Han@ibm.com>

Signed-off-by: Annabelle Huo <Annabelle.Huo@ibm.com>