Check if SCC is valid at startup
Currently the check to see if the SCC is valid is done at the first
compilation. This is done on the load run to see if it is ok to load a
method that was stored to the SCC. However, other parts of the compiler,
such as the IProfiler, don't use this check.

This commit checks the validity of the SCC at startup and updates the
global flag that all parts of the compiler check. If a AOT header does
not exist at this point, a header will be created.

Signed-off-by: Irwin D'Souza <dsouzai.gh@gmail.com>