Fix '-XXjitdirectory' option on Linux on Z
On Linux on Z, `libj9jit` library dynamically links the `libj9zlib` library
available in OpenJ9. In regular world, both library sits in the same directory
and when it need to load `libj9zlib` which is required by `libj9jit` it finds
it from the previously set path. When we specify the `-XXjitdirectory` option
it overwrites the path to load dynamically linked libraries from and JVM fails
to start as it can not find `libj9zlib` needed by JIT library. In order to fix
this issue, we need to load `libj9zlib` library before it loads `libj9jit`.

Signed-off-by: simonameng <simonameng97@gmail.com>