Add a null check to handle internalFindClassUTF8 failure
It was assumed that internalFindClassUTF8 will throw an exception on
failure based upon the wording of J9_FINDCLASS_FLAG_THROW_ON_FAIL.

internalFindClassUTF8 only sets an exception. It does not throw an
exception. So, a null check is added to protect the code dependent on
exceptionClass if internalFindClassUTF8 fails and returns null for
exceptionClass.

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>