Avoid Escape Analysis for objects or arrays involving value types
A value type field of an object created by the NEW bytecode must be
initialized with the default value of the field.  Escape Analysis is
not yet prepared to perform that initialization, so for now such objects
must not be stack allocated.

Similarly, an array whose element type is a value type has its elements
initialized by the ANEWARRAY or MULTIANEWARRAY bytecode with the
default value of the type, but Escape Analysis is not yet set up to
perform such initialization, so such arrays must not be stack allocated
for now.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>