Align fields in the MessageBuffer on a 32-bit boundary
Some architectures don't allow access to variables if they are not
aligned on their natural boundary. This commit ensures that all fields
in the MessageBuffer are aligned on a 32-bit boundary by inserting
1-3 bytes of padding where necessary.
Alignment of 64-bit fields can be achieved if we place them first inside
the buffer or if we impose another padding between the DataDescriptor
and the payload. This will be done in a different PR.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>