Reorganize Message::MetaData to ensure aligned access
Message::MetaData contained a 64-bit field which represented the concatenation
of the version and configuration flags. We wanted to preserve the position of
the version info (which is coming immediately after the integer representing
the size of the message), but this posed aligned access issues.
Tte solution was to break the 64-bit field into two components: one
representing the version and the other one representing the configuration
flags.
We also wanted the first descriptor coming after the MetaData to be 64-bit
aligned, so we shrunk the size of `type` and `_numDataPoints` fields to
uint16_t.
Moreover we deleted the `_metaDataOffset` field because the metaData always
comes immediatelly after the size of the message.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>