Ensure 64-bit fields are 64-bit aligned in the MessageBuffer
In order to access 64-bit fields in the MessageBuffer directly with a
single load operation, they need to be 64-bit aligned (otherwise on
some platforms we may have a performance penaly and on others we may
encounter bus errors). Currently, all fields in the message buffer
are 32-bit aligned through padding after the payload.
To ensure 64-bit alignment we allow for an offset of the payload from
the end of the DataDescriptor. Thus the layout becomes:
```
DataDescriptor (2*4=8 bytes)
_dataOffset (0 or 4 bytes)
_payload
_padding (0-3 bytes)
```

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>